create-project: Use specific major versions of dependencies (#917)
To help prevent breakage when newer versions of the dependencies are
released, but the corresponding Neutrino preset hasn't yet been
updated to support them. Ideally we'd have some way of auto-generating
these, but for now this at least avoids future breakage.

I was hoping to pin the Neutrino monorepo packages too, but that
would require a significant refactor, so that will have to wait until
another time.

Fixes #911.