eslint: Set the eslint-plugin-react react version to latest (#1184)
Several of the `eslint-plugin-react` rules use the specified React
version to determine what warnings/errors to show (for example only
showing deprecation warnings for certain features if using a newer
version of React). Not setting a version, or setting one that's older
than the version of React being used means that valid warnings/errors
may not actually be shown.

Whilst `eslint-plugin-react` helpfully outputs a warning if an explicit
React version has not been specified, several popular presets (such
as `eslint-config-airbnb` and `eslint-config-standard-react`) set a
default version of `16.0`, which results in many rules being disabled.

As such, we're using the trick used by CRA, of setting the version to
`999.999.999`, which ends up being interpreted as "latest version"
by `eslint-plugin-react`'s rules.

Fixes #1155.